btn Borrar  y  atributo disabled

+++ scripts/main.js
@@ -12,89 +12,120 @@ let btnModificar = document.getElementById("btnPut");
 let btnEliminar = document.getElementById("btnDelete");
 let btnGuardar = document.getElementById("btnSendChanges");
 
-//
+// 
 
-let urlUsers = "https://6362c2c937f2167d6f6c7564.mockapi.io/users";
+let urlUsers = "https://6362c2c937f2167d6f6c7564.mockapi.io/users"
 // para get, post
-let urlID = "https://6362c2c937f2167d6f6c7564.mockapi.io/users/:id";
+let urlID = "https://6362c2c937f2167d6f6c7564.mockapi.io/users/:id"
 // para  get1, put, delete
 
-//funcion para enlistar
+
+//document.addEventListener(", () => {
+
+
+
 let lista = [];
 let result = document.getElementById("results");
 let listaHTML = "";
 function resultado(lista) {
-  lista.forEach((lista) => {
+  //let listaHTML = "";
+  lista.forEach(lista => {
+
     listaHTML += `
         <li>ID: ${lista.id}</li>
         <li>Nombre: ${lista.name}</li>
         <li>Apellido: ${lista.lastname}</li>
-        `;
+        `
     result.innerHTML = listaHTML;
-  });
-}
+  })
+};
 
-//funcion para mostrar los usuarios por id y la lista completa
+
 btnBuscar.addEventListener("click", () => {
-  let id = inputGetId.value;
+  let id = inputGetId.value
   if (id != "" || id > 0) {
+
     fetch("https://6362c2c937f2167d6f6c7564.mockapi.io/users/" + id)
       .then((response) => response.json())
       .then((data) => {
-        lista = data;
+        lista = data
         result.innerHTML = `
        <li>ID: ${lista.id}</li>
        <li>Nombre: ${lista.name}</li>
        <li>Apellido: ${lista.lastname}</li>
-       `;
-      });
-  } else {
+       `
+
+
+      })
+
+  }
+  else {
     fetch(urlUsers)
       .then((response) => response.json())
       .then((data) => {
         lista = data;
         console.log(data);
         resultado(lista);
+
       });
   }
-});
+})
 
-//funcion para agregar nuevo usuario
+
+
+
+
 btnAgregar.addEventListener("click", () => {
+
   let nombre = inputPostNombre.value;
   let apellido = inputPostApellido.value;
+  if (nombre != "" || apellido != "")
 
-  fetch("https://6362c2c937f2167d6f6c7564.mockapi.io/users/", {
-    headers: { "Content-Type": "application/json" },
-    method: "POST",
-    body: JSON.stringify({
-      name: nombre,
-      lastname: apellido,
-    }),
-  })
-    .then((response) => response.json())
-    .then((data) => {
-      lista = data;
-      resultado(lista);
-    });
-});
+    fetch("https://6362c2c937f2167d6f6c7564.mockapi.io/users/", {
 
-//modificar modal
-btnModificar.addEventListener("click", () => {
-  let modalNombre = document.getElementById("inputPutNombre");
-  let modalApellido = document.getElementById("inputPutApellido");
+      headers: { 'Content-Type': 'application/json' },
+      method: 'POST',
+      body: JSON.stringify({
+        name: nombre,
+        lastname: apellido,
+      })
+    })
 
-  let id = inputPutId.value;
-  if (id != "" || id > 0) {
-    fetch("https://6362c2c937f2167d6f6c7564.mockapi.io/users/" + id)
       .then((response) => response.json())
       .then((data) => {
-        lista = data;
-        modalApellido.value = lista.lastname;
-        modalNombre.value = lista.name;
-      });
+
+        lista = data
+        resultado(lista);
+
+
+      })
+
+
+})
+
+
+
+btnModificar.addEventListener("click", () => {
+
+  let modalNombre = document.getElementById("inputPutNombre")
+  let modalApellido = document.getElementById("inputPutApellido")
+
+  let id = inputPutId.value
+  if (id != "" || id > 0) {
+
+    fetch("https://6362c2c937f2167d6f6c7564.mockapi.io/users/" + id)
+
+      .then((response) => response.json())
+      .then((data) => {
+        lista = data
+        modalApellido.value = lista.lastname
+        modalNombre.value = lista.name
+
+
+      })
+
   }
-});
+})
 
 let modalNombre = document.getElementById("inputNombre");
 let modalApellido = document.getElementById("inputApellido");
@@ -115,6 +146,7 @@ btnModificar.addEventListener("click", () => {
       });
   }
 });
+
 
 btnGuardar.addEventListener("click", () => {
   let modalNombre = document.getElementById("inputPutNombre");
@@ -138,12 +170,56 @@ btnGuardar.addEventListener("click", () => {
     });
 });
 
-// para eliminar un usuario
+//eliminar usuario
 btnEliminar.addEventListener("click", () => {
-  let id = inputDelete.value;
-  fetch("https://6362c2c937f2167d6f6c7564.mockapi.io/users/" + id, {
-    method: "DELETE",
-  })
-    .then((res) => res.json())
-    .then((data) => {});
+  let id = inputDelete.value
+  if (id != "" || id > 0) {
+
+    fetch("https://6362c2c937f2167d6f6c7564.mockapi.io/users/" + id, {
+      method: 'DELETE',
+    })
+      .then(res => res.json())
+      .then(data => {
+
+        lista = data
+        resultado(lista);
+      });
+  }
+
+});
+
+
+
+//Botones disabled
+
+btnAgregar.disabled = true
+btnModificar.disabled = true
+btnEliminar.disabled = true
+inputPostNombre.addEventListener("change", () => {
+
+  if (inputPostNombre.value === "" && inputPostApellido.value === "") {
+    btnAgregar.disabled = true
+  } else {
+    btnAgregar.disabled = false
+  }
+
+})
+
+inputPutId.addEventListener("change", () => {
+  if (inputPutId.value === "") {
+    btnModificar.disabled = true
+  } else {
+    btnModificar.disabled = false
+  }
+
+})
+
+inputDelete.addEventListener("change", () => {
+  if (inputDelete.value === "") {
+    btnEliminar.disabled = true
+  } else {
+    btnEliminar.disabled = false
+
+  }
+
 });
